pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "localization-analyzer"
dynamic = ["version"]
description = "Professional localization analyzer for multi-platform projects"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sezgin Paksoy"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Localization",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
]
watch = ["watchdog>=3.0"]
progress = ["tqdm>=4.65"]

[project.scripts]
localization-analyzer = "localization_analyzer.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/localization-analyzer"
Documentation = "https://github.com/yourusername/localization-analyzer#readme"
Repository = "https://github.com/yourusername/localization-analyzer"
"Bug Tracker" = "https://github.com/yourusername/localization-analyzer/issues"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
